small fixes

+++ src/ft_vec2i.c
@@ -6,13 +6,13 @@
 /*   By: myli-pen <myli-pen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/18 14:32:13 by myli-pen          #+#    #+#             */
-/*   Updated: 2025/07/20 19:26:27 by myli-pen         ###   ########.fr       */
+/*   Updated: 2025/07/21 14:10:25 by myli-pen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft_math.h"
 
-t_vec2i	vec2i(float x, float y)
+t_vec2i	vec2i(int x, int y)
 {
 	t_vec2i	result;
